pyproject.toml
[tool.poetry]
name = "c-match"
version = "2.0.0"
description = "Проект match для консолидации данных и произведения матчинга контента"
authors = ["Rustam Aliev"]
license = "MIT"  # Добавьте лицензию, если она есть
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
alembic = "1.14.0"
sqlalchemy = "2.0.37"
fastapi = "0.115.7"
python-dotenv = "1.0.1"
asyncpg = "0.30.0"
greenlet = "3.1.1"
psycopg2-binary = "2.9.10"
uvicorn = "0.34.0"
